+++ libreria/src/main/java/ar/edu/untref/aydoo/Libreria.java
@@ -4,10 +4,10 @@ import java.util.ArrayList;
 
 public class Libreria {
 
-    public int calcularMontoDelMes(Cliente miCliente, int mes) {
+    public double calcularMontoDelMes(Cliente miCliente, int mes) {
         ArrayList<Compra> compasDelCliente = miCliente.getCuentaDelCliente().getComprasDelCliente(mes);
         if(compasDelCliente != null) {
-            int montoTotalDelMes = 0;
+            double montoTotalDelMes = 0;
             for (Compra compra:compasDelCliente) {
                 montoTotalDelMes = montoTotalDelMes + compra.getMontoTotalDeLaCompra();
             }
